requirements.txt
google-cloud-storage==1.42.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
Flask==2.0.1
